+++ src/Device/deviceSetup.ts
@@ -63,6 +63,8 @@ export interface IDeviceSetup {
 export interface DeviceSetup {
     deviceSetups: IDeviceSetup[];
     allowCustomDevice?: boolean; // allow custom J-Link device
+    confirmMessage?: string;
+    choiceMessage?: string;
 }
 
 export const prepareDevice =
@@ -80,7 +82,7 @@ export const prepareDevice =
             dispatch(closeDeviceSetupDialog());
         };
         const validDeviceSetups = deviceSetupConfig.deviceSetups.filter(
-            deviceSetups => deviceSetups.supportsProgrammingMode(device)
+            deviceSetup => deviceSetup.supportsProgrammingMode(device)
         );
 
         const possibleFirmware = validDeviceSetups
@@ -188,6 +190,7 @@ export const prepareDevice =
                             }
                         },
                         message:
+                            deviceSetupConfig.confirmMessage ??
                             'Device must be programmed, do you want to proceed?',
                     })
                 );
@@ -195,6 +198,7 @@ export const prepareDevice =
                 dispatch(
                     openDeviceSetupDialog({
                         message:
+                            deviceSetupConfig.confirmMessage ??
                             'Device must be programmed, do you want to proceed?',
                     })
                 );
@@ -210,7 +214,9 @@ export const prepareDevice =
                             proceedAction(index ?? 0);
                         }
                     },
-                    message: 'Which firmware do you want to program?',
+                    message:
+                        deviceSetupConfig.choiceMessage ??
+                        'Which firmware do you want to program?',
                 })
             );
         }